+++ src/components/Popover/index.stories.js
@@ -50,6 +50,41 @@ export const controlledPopover = (args) => {
   );
 };
 
+export const popoverWithinAPopover = (args) => (
+  <Popover triggerElement={<button>Open popover</button>}>
+    <div className="padding--y--s padding--x--m" style={{ overflow: 'hidden' }}>
+      <h3>Here is some content inside.</h3>
+      <p>This popover is overflow hidden.</p>
+      <p>
+        Here is something else to
+        <Popover triggerElement={<button>click</button>} {...args}>
+          <div className="padding--all--m">
+            <p>
+              You can see this modal even though the parent modal has its{' '}
+              <code>overflow</code> set to <code>hidden</code>.
+            </p>
+            <p>You can also click inside here without closing the parent modals.</p>
+          </div>
+        </Popover>
+      </p>
+    </div>
+  </Popover>
+);
+popoverWithinAPopover.args = {
+  isFixed: true,
+  appendTo: 'parent',
+};
+const hidden = { table: { disable: true } };
+popoverWithinAPopover.argTypes = {
+  triggerElement: hidden,
+  theme: hidden,
+  children: hidden,
+  placement: hidden,
+  isOpen: hidden,
+  onUserDismiss: hidden,
+  triggerType: hidden,
+};
+
 export default {
   component: Popover,
   title: 'components/Popover',